+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/TLChatParticipantAdmin.java
@@ -7,7 +7,9 @@ import java.io.InputStream;
 import java.io.OutputStream;
 
 import static com.github.badoualy.telegram.tl.StreamUtils.readInt;
+import static com.github.badoualy.telegram.tl.StreamUtils.readLong;
 import static com.github.badoualy.telegram.tl.StreamUtils.writeInt;
+import static com.github.badoualy.telegram.tl.StreamUtils.writeLong;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32;
 
@@ -28,7 +30,7 @@ public class TLChatParticipantAdmin extends TLAbsChatParticipant {
     public TLChatParticipantAdmin() {
     }
 
-    public TLChatParticipantAdmin(int userId, int inviterId, int date) {
+    public TLChatParticipantAdmin(long userId, int inviterId, int date) {
         this.userId = userId;
         this.inviterId = inviterId;
         this.date = date;
@@ -36,7 +38,7 @@ public class TLChatParticipantAdmin extends TLAbsChatParticipant {
 
     @Override
     public void serializeBody(OutputStream stream) throws IOException {
-        writeInt(userId, stream);
+        writeLong(userId, stream);
         writeInt(inviterId, stream);
         writeInt(date, stream);
     }
@@ -44,7 +46,7 @@ public class TLChatParticipantAdmin extends TLAbsChatParticipant {
     @Override
     @SuppressWarnings({"unchecked", "SimplifiableConditionalExpression"})
     public void deserializeBody(InputStream stream, TLContext context) throws IOException {
-        userId = readInt(stream);
+        userId = readLong(stream);
         inviterId = readInt(stream);
         date = readInt(stream);
     }
@@ -68,11 +70,11 @@ public class TLChatParticipantAdmin extends TLAbsChatParticipant {
         return CONSTRUCTOR_ID;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
