+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/TLPeerChannel.java
@@ -6,8 +6,8 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 
-import static com.github.badoualy.telegram.tl.StreamUtils.readInt;
-import static com.github.badoualy.telegram.tl.StreamUtils.writeInt;
+import static com.github.badoualy.telegram.tl.StreamUtils.readLong;
+import static com.github.badoualy.telegram.tl.StreamUtils.writeLong;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32;
 
@@ -19,26 +19,26 @@ public class TLPeerChannel extends TLAbsPeer {
 
     public static final int CONSTRUCTOR_ID = 0xbddde532;
 
-    protected int channelId;
+    protected long channelId;
 
     private final String _constructor = "peerChannel#bddde532";
 
     public TLPeerChannel() {
     }
 
-    public TLPeerChannel(int channelId) {
+    public TLPeerChannel(long channelId) {
         this.channelId = channelId;
     }
 
     @Override
     public void serializeBody(OutputStream stream) throws IOException {
-        writeInt(channelId, stream);
+        writeLong(channelId, stream);
     }
 
     @Override
     @SuppressWarnings({"unchecked", "SimplifiableConditionalExpression"})
     public void deserializeBody(InputStream stream, TLContext context) throws IOException {
-        channelId = readInt(stream);
+        channelId = readLong(stream);
     }
 
     @Override
@@ -58,11 +58,11 @@ public class TLPeerChannel extends TLAbsPeer {
         return CONSTRUCTOR_ID;
     }
 
-    public int getChannelId() {
+    public long getChannelId() {
         return channelId;
     }
 
-    public void setChannelId(int channelId) {
+    public void setChannelId(long channelId) {
         this.channelId = channelId;
     }
 }
